move some tests from actions to bundle
